refactor common functionality into utility module
